修改两次注册流程，规范化代码

+++ src/main/java/cn/telcom/enrol/dao/IBusinessDao.java
@@ -6,6 +6,6 @@ import cn.telcom.enrol.bean.Business;
  * @author kenny_peng
  * @created 2019/8/2 15:32
  */
-public interface IBusinessDao extends IBaseDao<Long, Business> {
+public interface IBusinessDao extends IBaseDao<Integer, Business> {
     Business selectByBuNo(String buNo);
 }
